requirements.txt
openai
google-generativeai
chromadb
sentence-transformers
langchain
PymuPDF
python-docx
pydantic
python-dotenv
numpy
fastapi
uvicorn
aiohttp